refactor: gate Apollo Federation v2 features behind EE build tag
Mirrors the //go:build ee || dev pattern Tyk Streams uses
(gateway/mw_streaming.go vs mw_streaming_ee.go) and applies it to the
just-shipped Apollo Federation v2 work in the GraphQL middleware:

- Production: extract the V3 schema-augmentation block from
  GraphQLMiddleware.Init into augmentFederationSchema. The CE stub in
  mw_graphql_federation.go returns the SDL unchanged and emits a Warnf
  pointing customers at the EE build when @key is declared. The real
  implementation in mw_graphql_federation_ee.go performs UDG
  federation.BuildFederationSchema and the @key-detected proxy-mode
  passthrough augmentation. The Init call site is build-tag-agnostic;
  the linker selects which body fires.

- Tests: move all federation/proxy-federation/federation+subscription
  tests (21 functions across UDGFederation_*, ProxyMode_FederationPassthrough_*,
  V3_Subscription_FederationSubgraph_TWS,
  V3_Subscription_ProxyMode_FederationPassthrough_TWS, and
  V2WithKeyDirective_LogsWarning) from mw_graphql_test.go into a new
  mw_graphql_federation_test.go that carries the //go:build ee || dev
  tag. Add the same tag to apidef/adapter/gqlengineadapter/enginev3/
  adapter_udg_federation_test.go so the entire 10-function adapter
  federation file is EE-only.

- Symmetric CE coverage: mw_graphql_federation_ce_test.go (CE-only)
  asserts that loading a UDG/V3 API with @key directives succeeds, the
  warning fires, and `{ _service { sdl } }` is rejected (no augmentation,
  so _service is not a Query field). This is the federation analogue of
  the streams 403 stub.

The CE-compilable infrastructure stays untouched: data-source adapters
(entities_datasource*.go, adapter_udg.go, adapter_proxy_only.go) are
dormant in CE because the schema is never augmented. Subscription
plumbing, variables-null normalization, and StreamingClient remain CE
since they are general V3 infrastructure rather than federation-specific.

The TestGraphQLMiddleware_AcceptsNullVariables test (rover compat —
not federation-specific) keeps its CE residence; its query body swaps
the federation-only `_service { sdl }` for an introspection
`__schema { queryType { name } }` so the variables-null parsing fix is
exercised through the full pipeline on a CE-compilable schema.

Verified:
- go vet ./... and go vet -tags=ee ./... both clean
- gofmt -l clean over all modified/new files
- CE: 9 GraphQL middleware tests pass; CE federation stub test passes
- EE: 29 GraphQL middleware tests pass (8 CE-shared + 21 federation);
  17 enginev3 adapter tests pass (10 federation + 7 non-federation)
