+++ apidef/adapter/gqlengineadapter/enginev3/adapter_udg_federation_test.go
@@ -1,3 +1,5 @@
+//go:build ee || dev
+
 package enginev3
 
 import (
